GGRC-1935 Fix bug with improperly update of Issue widget
Newly created Issue is not displayed in Issues tab if issue was created
via Related Issues tab on Assessment's Info pane.

As widgets on inactive tabs don't listen models changes Issue counts update
can be perfomed from add-issue-button component with the help
of GGRC.Utils.CurrentPage.initCounts method.

Additionally, on Issues tab activation it's neccessary to check the equality
of viewModel.pageInfo.total and GGRC.Utils.CurrentPage.getCounts() and update
view of tree-widget-container component in case of inequality. It is assumed
that removal of issues is possible only on the active Issues tab.