Fix editing of cycle task entries
The bug here was because named tuples don't support keys starting with
underscore. Since this method is used for getting an initial state of
the object on PUT requests and no names with underscores are sent there
we can safely remove those from the initial state named tuple. On the
second note the bug only occurs on Cycle task entries, which will be
removed soon and initial state is needed for issue tracker integration
which does not work with cycle task entries.